v2.0.3: Comprehensive audit — 36 fixes across 16 files
Security: empty @grant granting all permissions, @connect bypass via
malformed URL, GM_addElement case-insensitive sanitization, signing
trustKey prototype pollution, Monaco adapter source validation.

Crashes: ScriptStorage search/getByNamespace/duplicate null-safety,
toggleScript undefined enabled, context menu tab.url crash, GM_cookie
null result, _makeRuleId collision, requireCache wrong key, signing
CRLF and whitespace fixes.

Performance: import getAll() O(N*M) to cached, resources O(N^2) to
chunked, urlchange handler dedup, console capture cap, audio tab
cleanup.

Dashboard: heatmap innerHTML+= DOM destruction, standalone
revokeObjectURL timing, profiles urlWatcher reset, chains deep copy,
diff/linter Uint32Array, devtools HAR version.

374/374 tests green. background.js 16,673 lines.

+++ pages/dashboard-diff.js
@@ -308,7 +308,7 @@ const DiffTool = (() => {
     const n = a.length, m = b.length;
     // For very large files, use a simpler approach
     if (n * m > 5000000) return _computeSimpleDiff(a, b);
-    const dp = Array.from({ length: n + 1 }, () => new Uint16Array(m + 1));
+    const dp = Array.from({ length: n + 1 }, () => new Uint32Array(m + 1));
     for (let i = 1; i <= n; i++) {
       for (let j = 1; j <= m; j++) {
         dp[i][j] = a[i - 1] === b[j - 1] ? dp[i - 1][j - 1] + 1 : Math.max(dp[i - 1][j], dp[i][j - 1]);
